[Presentation] Replaced wrapped_ascension_material.dart with material_quantity_row.dart
