+++ lib/presentation/weapon/widgets/weapon_detail_ascension_materials_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:shiori/domain/models/models.dart';
 import 'package:shiori/generated/l10n.dart';
-import 'package:shiori/presentation/shared/images/wrapped_ascension_material.dart';
 import 'package:shiori/presentation/shared/item_description_detail.dart';
+import 'package:shiori/presentation/shared/material_quantity_row.dart';
 import 'package:shiori/presentation/shared/styles.dart';
 import 'package:shiori/presentation/shared/utils/size_utils.dart';
 
@@ -56,16 +56,7 @@ class WeaponDetailAscensionMaterialsCard extends StatelessWidget {
   }
 
   TableRow _buildStatProgressionRow(WeaponAscensionModel model, double size) {
-    final materials = model.materials
-        .map(
-          (m) => WrappedAscensionMaterial(
-            itemKey: m.key,
-            image: m.image,
-            quantity: m.requiredQuantity,
-            size: size,
-          ),
-        )
-        .toList();
+    final materials = model.materials.map((m) => MaterialQuantityRow.fromAscensionMaterial(item: m, size: size)).toList();
     return TableRow(
       children: [
         TableCell(
