+++ lib/presentation/weapon/widgets/weapon_detail_crafting_materials.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:shiori/domain/models/models.dart';
 import 'package:shiori/generated/l10n.dart';
-import 'package:shiori/presentation/shared/images/wrapped_ascension_material.dart';
 import 'package:shiori/presentation/shared/item_description_detail.dart';
+import 'package:shiori/presentation/shared/material_quantity_row.dart';
 import 'package:shiori/presentation/shared/utils/size_utils.dart';
 
 class WeaponCraftingMaterials extends StatelessWidget {
@@ -19,16 +19,7 @@ class WeaponCraftingMaterials extends StatelessWidget {
   Widget build(BuildContext context) {
     final s = S.of(context);
     final size = SizeUtils.getSizeForCircleImages(context);
-    final rows = materials
-        .map(
-          (m) => WrappedAscensionMaterial(
-            itemKey: m.key,
-            image: m.image,
-            quantity: m.requiredQuantity,
-            size: size * 1.2,
-          ),
-        )
-        .toList();
+    final rows = materials.map((m) => MaterialQuantityRow.fromAscensionMaterial(item: m, size: size * 1.2)).toList();
 
     final body = Wrap(
       alignment: WrapAlignment.center,
